jax-rs client package refactoring
