+++ jaxrs-service/src/main/java/com/quorum/tessera/client/ClientFactory.java
@@ -1,4 +1,4 @@
-package com.quorum.tessera.node;
+package com.quorum.tessera.client;
 
 import com.quorum.tessera.config.ServerConfig;
 import com.quorum.tessera.ssl.context.SSLContextFactory;
