+++ jaxrs-service/src/main/java/com/quorum/tessera/client/PostDelegate.java
@@ -1,7 +1,6 @@
-package com.quorum.tessera.node;
+package com.quorum.tessera.client;
 
 import com.quorum.tessera.api.model.ResendRequest;
-
 import javax.ws.rs.client.Client;
 import javax.ws.rs.client.Entity;
 import javax.ws.rs.core.MediaType;
