+++ jaxrs-service/src/main/java/com/quorum/tessera/client/RestP2pClient.java
@@ -3,7 +3,6 @@ package com.quorum.tessera.client;
 
 import com.quorum.tessera.api.model.ApiPath;
 import com.quorum.tessera.api.model.ResendRequest;
-import com.quorum.tessera.node.PostDelegate;
 import java.util.Objects;
 
 public class RestP2pClient implements P2pClient {
