+++ jaxrs-service/src/test/java/com/quorum/tessera/client/RestP2pClientTest.java
@@ -2,7 +2,6 @@ package com.quorum.tessera.client;
 
 import com.quorum.tessera.api.model.ApiPath;
 import com.quorum.tessera.api.model.ResendRequest;
-import com.quorum.tessera.node.PostDelegate;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
