+++ jaxrs-service/src/test/java/com/quorum/tessera/node/PostDelegateTest.java
@@ -2,6 +2,7 @@ package com.quorum.tessera.node;
 
 import com.quorum.tessera.api.model.ApiPath;
 import com.quorum.tessera.api.model.ResendRequest;
+import com.quorum.tessera.client.PostDelegate;
 import org.junit.Before;
 import org.junit.Test;
